requirements.txt
aiohttp==3.9.4
aiofiles==24.1.0
disnake==2.9.2
pillow==10.3.0
playwright==1.46.0
